delete some of the javascript test files and add those tests to the psat.js file. Reduces number of files in the project and makes the unit tests easier to find overall.
